+++ scripts/macros/classFeatures/paladin/oathOfDevotion/auraOfDevotion.js
@@ -24,7 +24,10 @@ export async function auraOfDevotion(token, selectedAura) {
         ],
         'flags': {
             'chris-premades': {
-                'aura': true
+                'aura': true,
+                'effect': {
+                    'noAnimation': true
+                }
             }
         }
     }